先redis秒杀；然后发送消息队列；最后redis逐条写入数据库

+++ jseckill-backend/src/main/java/com/liushaoming/jseckill/backend/entity/SuccessKilled.java
@@ -11,7 +11,7 @@ public class SuccessKilled {
 
     private long userPhone;
 
-    private short state;
+    private int state;
 
     private Date createTime;
 
@@ -34,11 +34,11 @@ public class SuccessKilled {
         this.userPhone = userPhone;
     }
 
-    public short getState() {
+    public int getState() {
         return state;
     }
 
-    public void setState(short state) {
+    public void setState(int state) {
         this.state = state;
     }
